pyproject.toml
[tool.poetry]
name = "photonbend"
version = "0.1.1"
description = "Photonbend allows one to convert photos between different sorts of lenses, rotate photos and make panoramas."
authors = ["Edson Moreira <w.moreirae@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
Pillow = "^9.0.1"
numpy = ">=1.18,<1.22"
numba = "^0.55.1"
scipy = "^1.8.0"
click = "^8.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"

[tool.poetry.scripts]
photonbend = "photonbend.scripts.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
